Refactor split files
