+++ main.c
@@ -6,143 +6,6 @@
 #include "color.h"
 #include "mlx.h"
 
-t_ray	*ray_new(pvector *start, pvector *direction)
-{
-	t_ray	*ray;
-
-	ray = calloc(1, sizeof(*ray));
-	ray->start = start;
-	ray->direction = direction;
-	return (ray);
-}
-
-t_intersection_point	*sphere_get_intersection(float t, t_ray *ray, t_sphere *sphere)
-{
-	pvector					*position;
-	t_intersection_point	*intersection;
-
-	position = pvector_add(ray->start, pvector_mul(ray->direction, t));
-	intersection = calloc(1, sizeof(*intersection));
-	intersection->distance = t;
-	intersection->position = position;
-	intersection->normal = pvector_sub(position, sphere->center);
-	pvector_normalize(intersection->normal);
-	return (intersection);
-}
-
-t_intersection_point	*sphere_test_intersection(t_sphere *sp, t_ray *ray)
-{
-	pvector *d = ray->direction;
-	pvector *s = ray->start;
-	pvector *pc = sp->center;
-	float	r = sp->radius;
-	float	a = pvector_dot(d, d);
-	float	b = 2.0 * pvector_dot(pvector_sub(s, pc), d);
-	float	c = pvector_magsq(pvector_sub(s, pc)) - r * r;
-	float	D = b * b - 4.0 * a * c;
-	float	t1, t2;
-
-	if (D > 0)
-	{
-		t1 = (-b - sqrt(D)) / (2.0 * a);
-		t2 = (-b + sqrt(D)) / (2.0 * a);
-		if (t1 > 0)
-			return (sphere_get_intersection(t1, ray, sp));
-		else if (t2 > 0)
-			return (sphere_get_intersection(t2, ray, sp));
-	}
-	else if (D == 0)
-	{
-		t1 = (-b) / (2.0 * a);
-		if (t1 > 0)
-			return (sphere_get_intersection(t1, ray, sp));
-	}
-	return (NULL);
-}
-
-t_intersection_point	*plane_get_intersection(float t, t_ray *ray, t_plane *plane)
-{
-	pvector					*position;
-	t_intersection_point	*intersection;
-
-	position = pvector_add(ray->start, pvector_mul(ray->direction, t));
-	intersection = calloc(1, sizeof(*intersection));
-	intersection->distance = t;
-	intersection->position = position;
-	intersection->normal = pvector_copy(plane->normal);
-	pvector_normalize(intersection->normal);
-	return (intersection);
-}
-
-t_intersection_point	*plane_test_intersection(t_plane *plane, t_ray *ray)
-{
-	pvector *d = ray->direction;
-	pvector *s = ray->start;
-	pvector *pc = plane->center;
-	float	t;
-	float	dndot;
-
-	dndot = pvector_dot(d, plane->normal);
-	if (dndot != 0)
-	{
-		t = pvector_dot(pvector_sub(pc, s), plane->normal) / dndot;
-		if (t > 0)
-			return (plane_get_intersection(t, ray, plane));
-	}
-	return (NULL);
-}
-
-t_intersection_point	*test_intersection(t_shape *shape, t_ray *ray)
-{
-	t_intersection_point	*intersection;
-
-	if (shape->kind == SPHERE)
-		intersection = sphere_test_intersection(shape, ray);
-	else if (shape->kind == PLANE)
-		intersection = plane_test_intersection(shape, ray);
-	else
-		exit(1);
-	return (intersection);
-}
-
-#include <float.h>
-t_intersection_test_result	*test_intersection_with_all(t_scene *scene, t_ray *ray)
-{
-	return (test_intersection_with_all_overload(scene, ray, FLT_MAX, false));
-}
-
-t_intersection_test_result	*test_intersection_with_all_overload(t_scene *scene, t_ray *ray, float max_dist, bool exit_once_found)
-{
-	t_intersection_test_result	*res;
-
-	res = NULL;
-	for (t_shape *shape = scene->shapes; shape; shape = shape->next)
-	{
-		t_intersection_point	*intersection;
-
-		intersection = test_intersection(shape, ray);
-		if (intersection == NULL)
-			continue ;
-		if (intersection->distance >= max_dist)
-			continue ;
-		if (res == NULL)
-			res = calloc(1, sizeof(*res));
-		if (res->intersection_point == NULL)
-		{
-			res->intersection_point = intersection;
-			res->shape = shape;
-		}
-		else if (intersection->distance < res->intersection_point->distance)
-		{
-			res->intersection_point = intersection;
-			res->shape = shape;
-		}
-		if (exit_once_found && res->intersection_point->distance < max_dist)
-			return (res);
-	}
-	return (res);
-}
-
 t_ray *get_ray(int x, int y, pvector *camera)
 {
 	pvector *x_dir = pvector_new(1, 0, 0);
@@ -153,329 +16,6 @@ t_ray *get_ray(int x, int y, pvector *camera)
 	pvector *ray_dir = pvector_sub(pvector_add(pvector_mul(x_dir, u), pvector_mul(y_dir, v)), camera);
 	pvector_normalize(ray_dir);
 	return (ray_new(camera, ray_dir));
-}
-
-#define C_EPSILON (1.0 / 512.0)
-t_ray	*get_shadow_ray(t_intersection_point *intersection, t_light_source *ls, float *distance_to_ls_loc)
-{
-	pvector	*ray_start;
-	pvector	*ray_dir;
-
-	ray_dir = pvector_sub(ls->position, intersection->position);
-	*distance_to_ls_loc = pvector_mag(ray_dir) - C_EPSILON;
-	pvector_normalize(ray_dir);
-	ray_start = pvector_add(intersection->position, pvector_mul(ray_dir, C_EPSILON));
-	return (ray_new(ray_start, ray_dir));
-}
-
-t_fcolor	*ambient_light(t_ray *ray, t_shape *shape)
-{
-	t_fcolor	*ambient_intensity;
-	
-	(void)ray;
-	ambient_intensity = fcolor_new(0.1, 0.1, 0.1);
-	return (fcolor_mul(ambient_intensity, shape->material->amibient_factor));
-}
-
-t_fcolor	*diffuse_light(t_shape *shape, t_intersection_point *intersection, t_lighting *lighting)
-{
-	t_fcolor	*color;
-	pvector	*n;
-	pvector	*l;
-	float	nldot;
-
-	n = intersection->normal;
-	l = lighting->direction;
-	nldot = constrain(pvector_dot(l, n), 0 , 1);
-	color = fcolor_mul(lighting->intencity, shape->material->diffuse_factor);
-	color = fcolor_mul(color, fcolor_new(nldot, nldot, nldot));
-	return (color);
-}
-
-t_fcolor	*specular_light(t_ray *ray, t_shape *shape, t_intersection_point *intersection, t_lighting *lighting)
-{
-	pvector	*reflection;
-	pvector *b;
-	pvector	*view;
-	float	nldot;
-	float	vrdot;
-	float	vrdot_pow_alpha;
-	t_fcolor	*ki;
-
-	nldot = pvector_dot(intersection->normal, lighting->direction);
-	b = pvector_mul(intersection->normal, nldot * 2);
-	reflection = pvector_sub(b, lighting->direction);
-	view = pvector_mul(ray->direction, -1);
-	pvector_normalize(view);
-	vrdot = pvector_dot(view, reflection);
-	if (nldot <= 0 || vrdot <= 0)
-		return (fcolor_new(0, 0, 0));
-	vrdot_pow_alpha = pow(vrdot, shape->material->shineness);
-	ki = fcolor_mul(shape->material->specular_factor, lighting->intencity);
-	return (fcolor_mul(ki, fcolor_new(vrdot_pow_alpha, vrdot_pow_alpha, vrdot_pow_alpha)));
-}
-
-t_shape	*shape_new(t_shape_kind kind)
-{
-	t_shape	*sp;
-
-	sp = calloc(1, sizeof(*sp));
-	sp->kind = kind;
-	return (sp);
-}
-
-t_light_source *light_source_new(t_light_kind kind)
-{
-	t_light_source *src;
-
-	src = calloc(1, sizeof(*src));
-	src->kind = kind;
-	return (src);
-}
-
-t_lighting *lighting_at(pvector *pos, t_light_source *light_source)
-{
-	t_lighting	*lighting;
-
-	lighting = calloc(1, sizeof(*lighting));
-	lighting->direction = pvector_sub(light_source->position, pos);
-	lighting->distance = pvector_mag(lighting->direction);
-	pvector_normalize(lighting->direction);
-	lighting->intencity = light_source->intencity;
-	return (lighting);
-}
-
-t_material	*get_default_material(t_fcolor	*color)
-{
-	t_material		*material;
-
-	material = material_new();
-	material->amibient_factor = fcolor_new(0.01, 0.01, 0.01);
-	material->diffuse_factor = color;
-	material->specular_factor = fcolor_new(0.30, 0.30, 0.30);
-	material->shineness = 8.0;
-	material->use_perfect_reflectance = false;
-	material->catadioptric_factor = fcolor_new(0, 0, 0);
-	return (material);
-}
-
-t_shape	*get_shapes(void)
-{
-	t_shape		head;
-	t_shape		*shape;
-	
-	head.next = NULL;
-	shape = &head;
-	{
-		shape = shape->next = shape_new(SPHERE);
-		shape->center = pvector_new(3, 0, 25);
-		shape->radius = 1.0;
-		shape->material = get_default_material(fcolor_new(0.69, 0, 0));
-	}
-	{
-		shape = shape->next = shape_new(SPHERE);
-		shape->center = pvector_new(2, 0, 20);
-		shape->radius = 1.0;
-		shape->material = get_default_material(fcolor_new(0, 0.69, 0));
-	}
-	{
-		shape = shape->next = shape_new(SPHERE);
-		shape->center = pvector_new(1, 0, 15);
-		shape->radius = 1.0;
-		shape->material = get_default_material(fcolor_new(0, 0, 0.69));
-	}
-	{
-		shape = shape->next = shape_new(SPHERE);
-		shape->center = pvector_new(0, 0, 10);
-		shape->radius = 1.0;
-		shape->material = get_default_material(fcolor_new(0, 0.69, 0.69));
-	}
-	{
-		shape = shape->next = shape_new(SPHERE);
-		shape->center = pvector_new(-1, 0, 5);
-		shape->radius = 1.0;
-		shape->material = get_default_material(fcolor_new(0.69, 0, 0.69));
-	}
-	{
-		shape = shape->next = shape_new(PLANE);
-		shape->center = pvector_new(0, -1, 0);
-		shape->normal = pvector_new(0, 1, 0);
-		shape->material = get_default_material(fcolor_new(0.69, 0.69, 0.69));
-	}
-	return (head.next);
-}
-
-#define MAX_RECURSION 8
-
-t_fcolor	*ray_trace_recursive(t_scene *scene, t_ray *ray, int recursion_level)
-{
-	t_fcolor					*R;
-	t_intersection_test_result	*res;
-	t_lighting					*lighting;
-
-	if (recursion_level > MAX_RECURSION)
-		return (NULL);
-	res = test_intersection_with_all(scene, ray);
-	if (res)
-	{
-		R = ambient_light(ray, res->shape);
-		for (t_light_source *ls = scene->light_sources; ls; ls = ls->next)
-		{
-			t_ray	*shadow_ray;
-			float	distance_to_ls;
-
-			shadow_ray = get_shadow_ray(res->intersection_point, ls, &distance_to_ls);
-			if (test_intersection_with_all_overload(scene, shadow_ray, distance_to_ls, true))
-				continue ;
-			lighting = lighting_at(res->intersection_point->position, ls);
-			R = fcolor_add(R, diffuse_light(res->shape, res->intersection_point, lighting));
-			R = fcolor_add(R, specular_light(ray, res->shape,res->intersection_point, lighting));
-		}
-		if (res->shape->material->use_perfect_reflectance)
-		{
-			t_fcolor	*R_re, *R_mirror;
-			t_ray						*re_ray;
-			pvector	*p, *n, *v, *re, *re_ray_start;
-			float	vndot;
-
-			p = res->intersection_point->position;
-			n = res->intersection_point->normal;
-			v = pvector_mul(ray->direction, -1);
-			pvector_normalize(v);
-			vndot = pvector_dot(v, n);
-			re = pvector_sub(pvector_mul(n, 2 * vndot), v);
-			pvector_normalize(re);
-			re_ray_start = pvector_add(p, pvector_mul(re, C_EPSILON));
-			re_ray = ray_new(re_ray_start, re);
-			R_re = ray_trace_recursive(scene, re_ray, recursion_level + 1);
-			if (R_re)
-			{
-				R_mirror = fcolor_mul(res->shape->material->catadioptric_factor, R_re);
-				R = fcolor_add(R, R_mirror);
-			}
-		}
-		return (R);
-	}
-	else
-		return (fcolor_new(100.0 / 255.0, 149.0 / 255.0, 237.0 / 255.0)); // background color
-	return (R);
-}
-
-t_fcolor	*ray_trace(t_scene *scene, t_ray *ray)
-{
-	return (ray_trace_recursive(scene, ray, 0));
-}
-
-t_light_source	*get_light_sources(void)
-{
-	t_light_source	head;
-	t_light_source	*ls;
-
-	head.next = NULL;
-	ls = &head;
-	{
-		ls = ls->next = light_source_new(POINT);
-		ls->position = pvector_new(-5, 5, -5);
-		ls->intencity = fcolor_new(0.5, 0.5, 0.5);
-	}
-	{
-		ls = ls->next = light_source_new(POINT);
-		ls->position = pvector_new(5, 0, -5);
-		ls->intencity = fcolor_new(0.5, 0.5, 0.5);
-	}
-	{
-		ls = ls->next = light_source_new(POINT);
-		ls->position = pvector_new(5, 20, -5);
-		ls->intencity = fcolor_new(0.5, 0.5, 0.5);
-	}
-	return (head.next);
-}
-
-t_scene	*get_scene(void)
-{
-	t_scene	*scene;
-
-	scene = calloc(1, sizeof(*scene));
-	scene->shapes = get_shapes();
-	scene->light_sources = get_light_sources();
-	return (scene);
-}
-
-t_shape	*get_shapes2(void)
-{
-	t_shape		head;
-	t_shape		*shape;
-	
-	head.next = NULL;
-	shape = &head;
-	{
-		shape = shape->next = shape_new(SPHERE);
-		shape->center = pvector_new(-0.25, -0.5, 3.0);
-		shape->radius = 0.5;
-		shape->material = get_default_material(fcolor_new(0, 0, 0));
-		shape->material->amibient_factor = fcolor_new(0, 0, 0);
-		shape->material->diffuse_factor = fcolor_new(0, 0, 0);
-		shape->material->specular_factor = fcolor_new(0, 0, 0);
-		shape->material->shineness = 0;
-		shape->material->use_perfect_reflectance = true;
-		shape->material->catadioptric_factor = fcolor_new(1.0, 1.0, 1.0);
-	}
-	{
-		shape = shape->next = shape_new(PLANE);
-		shape->normal = pvector_new(0, 1, 0);
-		shape->center = pvector_new(0, -1, 0);
-		shape->material = get_default_material(fcolor_new(1, 1, 1));
-	}
-	{
-		shape = shape->next = shape_new(PLANE);
-		shape->normal = pvector_new(0, -1, 0);
-		shape->center = pvector_new(0, 1, 0);
-		shape->material = get_default_material(fcolor_new(1, 1, 1));
-	}
-	{
-		shape = shape->next = shape_new(PLANE);
-		shape->normal = pvector_new(-1, 0, 0);
-		shape->center = pvector_new(1, 0, 0);
-		shape->material = get_default_material(fcolor_new(0, 1, 0));
-	}
-	{
-		shape = shape->next = shape_new(PLANE);
-		shape->normal = pvector_new(1, 0, 0);
-		shape->center = pvector_new(-1, 0, 0);
-		shape->material = get_default_material(fcolor_new(1, 0, 0));
-	}
-	{
-		shape = shape->next = shape_new(PLANE);
-		shape->normal = pvector_new(0, 0, -1);
-		shape->center = pvector_new(0, 0, 5);
-		shape->material = get_default_material(fcolor_new(1, 1, 1));
-	}
-	return (head.next);
-}
-
-t_light_source	*get_light_sources2(void)
-{
-	t_light_source	head;
-	t_light_source	*ls;
-
-	head.next = NULL;
-	ls = &head;
-	{
-		ls = ls->next = light_source_new(POINT);
-		ls->position = pvector_new(0, 0.9, 2.5);
-		ls->intencity = fcolor_new(1.0, 1.0, 1.0);
-	}
-	return (head.next);
-}
-
-t_scene	*get_scene2(void)
-{
-	t_scene	*scene;
-
-	scene = calloc(1, sizeof(*scene));
-	scene->shapes = get_shapes2();
-	scene->light_sources = get_light_sources2();
-	return (scene);
 }
 
 int	main(void)
